requirements.txt
python-dotenv
django
djangorestframework
django-cors-headers
mysqlclient
coreapi